legacy/python/pyproject.toml
[project]
name = "new-ime"
version = "0.1.0"
description = "CTC-NAT based Japanese kana-kanji conversion engine"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "torch>=2.2",
    "transformers>=4.40",
    "tokenizers",
    "sentencepiece",
    "mecab-python3",
    "unidic-lite",
    "jaconv",
    "wandb",
    "accelerate",
    "onnxruntime",
    "protobuf>=4.0",
    "zstandard>=0.25.0",
    "ja-ginza>=5.2.0",
    "ginza>=5.2.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "llama-cpp-python",
    "huggingface-hub",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
